Fix clippy warnings and paths in props/mods.rs
