+++ packages/core-macro/src/component_body/mod.rs
@@ -143,109 +143,112 @@
 //!     }
 //! ```
 
+pub mod utils;
+pub use utils::is_type_eq;
+pub use utils::DeserializerArgs;
+pub use utils::DeserializerOutput;
+pub use utils::TypeHelper;
+
 use dioxus_core::{Element, Scope};
-use quote::ToTokens;
-use syn::{parse_quote, Path, Type};
+use syn::parse::{Parse, ParseStream};
+use syn::spanned::Spanned;
+use syn::*;
 
-mod component_body;
-pub use crate::component_body::component_body::ComponentBody;
-
-/// The output produced by a deserializer.
+/// General struct for parsing a component body.
+/// However, because it's ambiguous, it does not implement [`ToTokens`](quote::to_tokens::ToTokens).
 ///
-/// # For implementors
-/// Struct field guidelines:
-/// * Must be public, so that other deserializers can utilize them.
-/// * Should usually be [`Item`]s that you then simply combine in a [`quote!`]
-/// in the [`ComponentBodyDeserializer::output_to_token_stream2`] function.
-/// * If an [`Item`] might not be included, wrap it in an [`Option`].
-/// * Must be at the component function "level"/"context".
-/// For example, the [`InlinePropsDeserializer`](crate::component_body_deserializers::inline_props::InlinePropsDeserializer)
-/// produces two [`Item`]s; the function but with arguments turned into props, and the props struct.
-/// It does not return any [`Item`]s inside the struct or function.
-pub trait DeserializerOutput: ToTokens {}
+/// Refer to the [module documentation](crate::component_body) for more.
+pub struct ComponentBody {
+    /// The component function definition. You can parse this back into a [`ComponentBody`].
+    /// For example, you might modify it, parse it into a [`ComponentBody`], and deserialize that
+    /// using some deserializer. This is how deserializers use other deserializers.
+    ///
+    /// **`item_fn.sig.inputs` includes the context argument!**
+    /// Keep this in mind when creating deserializers, because you often might want to ignore it.
+    /// That might be annoying, but it would be bad design for this kind of struct to not be parsable from itself.
+    pub item_fn: ItemFn,
+    /// Parsing tries to ensure that this argument will be a [`Scope`].
+    /// **However, macros have limitations that prevent this from always working,
+    /// so don't take this for granted!**
+    pub cx_arg: FnArg,
+    /// The pattern (name) and type of the context argument.
+    pub cx_pat_type: PatType,
+    /// If the function has any arguments other than the context.
+    pub has_extra_args: bool,
+}
 
-/// The args passed to a [`ComponentBody`] when deserializing it.
-///
-/// It's also the struct that does the deserializing.
-/// It's called "DeserializerArgs", not "Deserializer". Why?
-/// Because "args" makes more sense to the caller of [`ComponentBody::deserialize`], which
-/// takes an [`DeserializerArgs`] argument. However, you can think of "DeserializerArgs" as the deserializer.
-pub trait DeserializerArgs<TOutput>: Clone
-where
-    TOutput: DeserializerOutput,
-{
+impl ComponentBody {
     // There's a lot of Results out there... let's make sure that this is a syn::Result.
     // Let's also make sure there's not a warning.
-    /// Creates a [`DeserializerOutput`] from the `self` args and a [`ComponentBody`].
-    /// The [`ComponentBody::deserialize`] provides a cleaner way of calling this function.
-    /// As a result, don't make this public when you implement it.
     #[allow(unused_qualifications)]
-    fn to_output(&self, component_body: &ComponentBody) -> syn::Result<TOutput>;
-}
-
-pub(crate) trait TypeHelper {
-    fn get_path() -> Path;
-    fn get_path_string() -> String {
-        Self::get_path().to_token_stream().to_string()
+    pub fn deserialize<TOutput, TArgs>(&self, args: TArgs) -> syn::Result<TOutput>
+    where
+        TOutput: DeserializerOutput,
+        TArgs: DeserializerArgs<TOutput>,
+    {
+        args.to_output(self)
     }
 }
 
-impl<'a> TypeHelper for Scope<'a> {
-    fn get_path() -> Path {
-        parse_quote!(::dioxus::core::Scope)
-    }
-}
+impl Parse for ComponentBody {
+    fn parse(input: ParseStream) -> Result<Self> {
+        let item_fn: ItemFn = input.parse()?;
+        let scope_type_path = Scope::get_path_string();
 
-impl<'a> TypeHelper for Element<'a> {
-    fn get_path() -> Path {
-        parse_quote!(::dioxus::core::Element)
-    }
-}
+        let (cx_arg, cx_pat_type) = if let Some(first_arg) = item_fn.sig.inputs.first() {
+            let incorrect_first_arg_err = Err(Error::new(
+                first_arg.span(),
+                format!("First argument must be a <{}>", scope_type_path),
+            ));
 
-/// Checks if a given generic that implements [`TypeHelper`] is equal to a [`Type`].
-///
-/// Returns `true` for both an imported and fully qualified [`TypeHelper::get_path`],
-/// but not "partially" qualified, such as: `use dioxus::core; core::Scope`.
-///
-/// # Edge cases
-/// These are caused by macro limitations.
-/// It's impossible to tell what module a type is imported from and if that type is an alias.
-/// Examples of some edge cases:
-/// * `is_type_eq<dioxus_core::Scope>(t) == true` where `t == type Scope = u8;` (an alias).
-/// * `is_type_eq<dioxus_core::Scope>(t) == false` where `t == core::Scope` from `use dioxus::core;`
-/// - It's possible to return `true` here, but `core` is too ambiguous,
-/// and as mentioned, it's impossible to know if a `core` belongs to `dioxus` or something else.
-fn is_type_eq<T>(input: &Type) -> bool
-where
-    T: TypeHelper,
-{
-    let scope_path = T::get_path();
-    let input_path: Path = parse_quote!(#input);
-    let scope_path_segs = scope_path.segments;
-    let input_path_segs = input_path.segments;
-    let input_seg_len = input_path_segs.len();
-
-    return if input_seg_len > 1 {
-        if input_seg_len == scope_path_segs.len() {
-            for i in 0..input_seg_len {
-                let input_seg = &input_path_segs[i];
-                let scope_seg = &scope_path_segs[i];
-
-                if input_seg.ident != scope_seg.ident {
-                    return false;
+            match first_arg.to_owned() {
+                FnArg::Receiver(_) => {
+                    return incorrect_first_arg_err;
+                }
+                FnArg::Typed(f) => {
+                    if is_type_eq::<Scope>(&f.ty) {
+                        (first_arg.to_owned(), f)
+                    } else {
+                        return incorrect_first_arg_err;
+                    }
                 }
             }
-
-            true
         } else {
-            false
-        }
-    } else if input_seg_len == 1 {
-        let scope_last_seg = scope_path_segs.last().unwrap();
-        let input_seg = input_path_segs.first().unwrap();
+            return Err(Error::new(
+                item_fn.sig.ident.span(), // Or maybe just item_f.sig.span()?
+                format!(
+                    "Must have at least one argument that's a <{}>",
+                    scope_type_path
+                ),
+            ));
+        };
 
-        scope_last_seg.ident == input_seg.ident
-    } else {
-        false
-    };
+        let element_type_path = Element::get_path_string();
+
+        match &item_fn.sig.output {
+            ReturnType::Default => {
+                return Err(Error::new(
+                    item_fn.sig.output.span(),
+                    format!("Must return a <{}>", element_type_path),
+                ))
+            }
+            ReturnType::Type(_, return_type) => {
+                if !is_type_eq::<Element>(return_type) {
+                    return Err(Error::new(
+                        item_fn.sig.output.span(),
+                        format!("Must return a <{}>", element_type_path),
+                    ));
+                }
+            }
+        }
+
+        let has_extra_args = item_fn.sig.inputs.len() > 1;
+
+        Ok(Self {
+            item_fn,
+            cx_arg,
+            cx_pat_type,
+            has_extra_args,
+        })
+    }
 }
